+++ persistent/ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## 2.14.0.0 (unreleased)
 
+* [#1343](https://github.com/yesodweb/persistent/pull/1343)
+    * Implement Type Literal based field definitions
 * [#1387](https://github.com/yesodweb/persistent/pull/1387)
     * Better UX with `insert`. We now report a type error when you try to
       `insert` an `Entity` or a function, and we also forbid `insert`ing if the
@@ -47,7 +49,6 @@
     * Increasing test coverage for errors thrown when parsing entity definitions
 
 ## 2.13.3.4
-
 * [#1379](https://github.com/yesodweb/persistent/pull/1379)
     * `mkPersist` now generates code that compiles under `NoFieldSelectors` and `DuplicateRecordFields` even if field labels are not prefixed
 * [#1376](https://github.com/yesodweb/persistent/pull/1376)